Add test jwks keyset utility
Add simple utility to generate static jwks.KeySet with prepopulated
keys. This is starting to be used in the app and rdk allover the place.
